requirements.txt
osmium
haversine
utm
pandas
pillow
numpy